+++ e2e/playwright.config.js
@@ -34,8 +34,6 @@ export default defineConfig({
   /* Shared settings for all the projects below. See https://playwright.dev/docs/api/class-testoptions. */
   use: {
     actionTimeout: 30 * 1000,
-    /* Base URL to use in actions like `await page.goto('/')`. */
-    baseURL: process.env.E2E_TESTS_ENV_URL ?? 'https://hds.hel.fi',
     ignoreHTTPSErrors: true,
     /* Collect trace when retrying the failed test. See https://playwright.dev/docs/trace-viewer */
     trace: 'on-first-retry',
@@ -43,6 +41,14 @@ export default defineConfig({
     // contextOptions: { recordVideo: { dir: './report/videos/' } },
     screenshot: 'only-on-failure',
     video: 'off',
+    channel: 'chromium',
+    launchOptions: {
+      slowMo: 100,
+    },
+    retries: 3,
+    expect: {
+      timeout: 2000,
+    },
   },
   projects: [
     {
@@ -51,7 +57,7 @@ export default defineConfig({
       use: {
         ...devices['Desktop Chrome'],
         viewport: { width: 1280, height: 720 },
-        isMobile: false,
+        hasTouch: false,
       },
     },
     {
@@ -59,8 +65,8 @@ export default defineConfig({
       testMatch: [/tests/],
       use: {
         ...devices['Desktop Chrome'],
-        viewport: { width: 320, height: 576 },
-        isMobile: true,
+        viewport: { width: 320, height: 720 },
+        hasTouch: true,
       },
     },
   ],